doing some work for enumerating how the accounts work together
